perf(size): Mark `error_ctx_*` as noreturn
Add the `noreturn` marker to the `error_ctx_*` family of functions in
'eval_cps.c'. This potentially helps the compiler do optimizations.

It actually did decrease the size of the compiled binary slightly,
the repl ELF compiled on my system went from 1532036 bytes to 1532028
bytes, a whole 8 bytes less. (Massive improvement!)

Note that this change means that a lot of redundant returns which are
present to make static analysis tools happy could be removed. But those
returns had comments which talked about the static analysis Infer
written by FaceBook, and since I haven't bothered trying to install it I
decided to not remove them as I wouldn't have been able to confirm that
Infer actually understood the `noreturn` markers.