docs: Reorganize to define special forms more clearly
Moves around some of the functions in "lbmref.lisp" to other extensions,
and adds new "Special form" tags to all special forms, where that status
is defined by being in the `evaluators` list in "eval_cps.c".

I also tweaked the phrasing and styling in a few places where I happened
to spot improvement opportunities.

Also made it possible to override the `lbm` binary used in the doc's
Makefile.