feat: Make `apply` a built-in
Transitioned `apply` to be a built-in and change it's behavior to not
evaluate it's arguments. Moves its documentation from dynref.lisp to
lbmref.lisp, and update it to explain the new behavior.

Unfortunately this essentially required duplicating all branches in the
`APPLICATION_START` continuation, implementing each kind of applicable
value. So the binary size increased by a relatively high amount. When I
compile the stormcore_100s target in the bldc repository with the
changes from this commit the size went from 449284 to 450028 bytes, an
increase of 744.