perf: Make `setup_*` always inline
Adds the inline and GCC always_inline attributes to the
`setup_cont(_sp)` and `setup_macro` functions added in the last commit.
This should in theory make them faster since we avoid the function call.
This of course also increases the binary size quite significantly,
from 450028 to 450444 bytes, an increase of 416 bytes.

It's interesting to note that when I only added the inline attributes
(leaving out the always_inline) and analyzed the resulting assembly
after building bldc, `setup_cont` and `setup_cont_sp` were only partly
inlined. The function labels were still present in the output assembly,
but they don't contain the entire function's logic. I also *think* that
the part which is in the separate function is only called on the error
path, but I'm not experienced enough in assembly to tell for sure.

Either way, in this case the resulting binary size was 450428, meaning
that adding the always_inline attribute only made the result 16 bytes
larger than that. It also most likely has slightly better performance,
even if only for the error paths.