Add `in` fundamental
Adds the `in` fundamental function which checks if a given value is
included in a list. A good use case for it is to check if a value is in
some constant set of values. This would traditionally be achieved as a
series of `eq` expressions wrapped in an`or` expression, e.g.
`(or (eq value 'a) (eq value 'b) (eq value 'c))`. With `in` this can be
written as `(in value '(a b c)')`, which is arguably much easier to
parse (at least after you've become familiar with the pattern).

The reason this should be a builtin is that it's much more efficient
than an equivalent user implementation. This is the best I could think
of:
```clj
(defun in-user (value lst)
  (foldl
    (fn (result lst-value)
      (or result (eq lst-value value)))
    false
    lst))
```

Then comparing `(in-user 3 '(1 2 3)')` to `(in 3 '(1 2 3)')` the builtin
one was 5x faster! A nice bonus is that it's also slightly more
efficient than manually unrolling the equality checks inside an `or`
expression (it was just 1.3x faster though). This was measured by
running each expression 10000 times, taking the average. The tests were
ran in the 32-bit repl on my x86 machine. As it's not an embedded system
the results should of course be taken with a grain of salt.