Restore `setassoc` behavior for missing keys.
Previously, if `setassoc` was called with a key which wasn't set it
would return a list with an added new pair. This behavior was removed in
3c375f7, with it now instead raising an eval_error. This commit
reimplements the old behavior when it comes to missing keys.

While the behavior was never explicitly mentioned in the documentation
it is still a useful feature, that is hard to emulate in an efficient
manner, since a userland implementation would have to check if the keys
exist before inserting them, which isn't implemented in any built-in
function.

We also have production code at Lind which depend on this behavior, so
if anything it would be helpful to not break backward compatability.