Add support for binding built-in symbols
Adds support for binding arbitrary symbols to the global and local
environments, regardless of if they're built-in or not. Also updates the
documentation to reflect the new behavior.

This allows for fun code like the following ;)
```clj
(defun list ()
    'im-not-a-list
)

(def t 'nil)
(def nil 't)

(defun lerp (a b t) (+ (* (- 1 t) a) (* t b)))

(print (list 1 2 3))
; 'im-not-a-list
(print (if t 'yes 'no))
; 'no
(print (lerp 0 2 0.4))
; 0.800000f32
```

But the more practical reason to desire this is to allow new built-ins
and extensions to not be a breaking change for programs which happened
to be using those symbols.

Unfortunately, we should expect a performance hit, as now all built-in
symbols need to check the environments first. We'll need to measure the
actual impact to determine if it's acceptable.