Fix overflowing quota timer preventing ctx switch
If the active quota timer overflows (meaning it's close to 0) while the
active quota timestamp hasn't (meaning it's close to `UINT32_MAX`), then
the currently active context will continue running for a very long time,
since the current timestamp has to be larger than the quota timestamp
for the contexts to be switched. The runtime may break out of this state
after the current timestamp has gone all the way up to the edge of
overflowing (this takes ~71 minutes). However, this isn't guaranteed, as
the current timestamp may skip over the quota and overflow within a
single evaluation step.

This commit handles this scenario by instead comparing if the difference
between the current and quota timestamps is negative. The overflow then
works out so that, for the above scenario, the current timestamp is
correctly considered as "ahead of" the quota timestamp, causing the
context to be switched.

Note that the implementation uses unsigned integers, so when it checks
if the difference is "negative", it's really checking if it's larger
than the max singed 32-bit int (or in other words, half of the max
unsigned 32-bit int). It has to be done this way since causing overflow
with signed integers is undefined behavior in C, and there doesn't seem
to be any way to accomplish what we want in a well defined manor. :(

Luckily, this fix shouldn't have that big of a performance impact, since
it only adds an additional subtraction compared to the previous code.