repl: Only print context result if started by REPL
Makes it so that the REPL only prints the result of a context if it was
directly started by the REPL, i.e. if it was created by the user
pressing enter.

For example, this is the what would be printed if you spawned a context
before this change:
```
# 'result
> result
# (spawn (fn () 'thd-result))
> 881
> thd-result
```
With the change, this is output:
```
# 'result
> result
# (spawn (fn () 'thd-result))
> 881
```
Note how the result `thd-result` from the spawned context isn't printed.