repl: Redraw prompt after anything is printed
Makes the prompt be cleared before anything is printed, and have it be
redrawn afterwards below.
This is done by introducing a new printf wrapper, `printf_redraw_prompt`
which must be used instead of `printf` for all threads other than the
main one.

It's exact behavior is that it first checks if the prompt was the last
thing printed, and if so calls a readline function to clear the prompt
(note: this function handles multiline prompts correctly ^^). It then
prints the output, and finally, if the printed text ended in a newline,
it redraws the prompt. This makes sure that the prompt is only drawn
at the start of a line, which is necessary to avoid other printed output
from being cleared, as clearing the prompt also clears any other text
that was on the line the prompt was drawn to.

I had to make the `printf_callback` variable in 'eval_cps.c' global,
since the repl_exts need to be able to access the safe printf version.
I also renamed it to `lbm_printf_callback`.